fix(activerecord): address Copilot review #3 — extract buildColumnSerializer, fix comment
Comment 1: buildColumnSerializer was duplicated (inlined in store.ts,
private in serialization.ts). Extracted to coders/build-column-serializer.ts
which only imports from coders/ — no type/json.ts dep — so both store.ts
and attribute-methods/serialization.ts can import it without circular refs.

Comment 2: storeAccessorFor comment said "registered by Base.store()" but
setStoreCoder is called by standalone store() too. Fixed to say "store()".
