+++ packages/activerecord/src/attribute-methods/serialization.ts
@@ -12,7 +12,6 @@
  *
  * Mirrors: ActiveRecord::AttributeMethods::Serialization
  */
-import { ColumnSerializer as CodersColumnSerializer } from "../coders/column-serializer.js";
 import { JSON as CodersJSON } from "../coders/json.js";
 import { Json as JsonType } from "../type/json.js";
 export interface Serialization {
@@ -55,29 +54,6 @@ export class ColumnSerializer {
   load(raw: unknown): unknown {
     return this.coder.load(raw);
   }
-}
-
-type CoderLike = { dump(v: unknown): string; load(v: unknown): unknown };
-
-/** @internal */
-function buildColumnSerializer(
-  attrName: string,
-  coder: unknown,
-  type: unknown,
-  _yaml?: Record<string, unknown>,
-): unknown {
-  const resolvedCoder = coder === globalThis.JSON ? CodersJSON : coder;
-
-  // coder.respond_to?(:new) && !coder.respond_to?(:load) → instantiate as constructor
-  if (typeof resolvedCoder === "function" && !("load" in resolvedCoder)) {
-    return new (resolvedCoder as any)(attrName, type);
-  }
-
-  if (type && type !== Object) {
-    return new CodersColumnSerializer(attrName, resolvedCoder as CoderLike, type as any);
-  }
-
-  return resolvedCoder;
 }
 
 /** @internal */
